Changed field types

+++ app/utils/payment.go
@@ -1,9 +1,9 @@
 package utils
 
 type PaymentSettings []struct {
-	Amount   int `json:"amount"`
-	DataMb   int `json:"data_mb"`
-	TimeMins int `json:"time_mins"`
+	Amount   float64 `json:"amount"`
+	DataMb   float64 `json:"data_mb"`
+	TimeMins float64 `json:"time_mins"`
 }
 
 var DefaultPaymentSettings = PaymentSettings{}
